requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
litellm==1.67.5
MarkupSafe==3.0.2
mcp==1.7.1
multidict==6.4.3
openai==1.77.0
openai-agents==0.0.14
packaging==25.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zipp==3.21.0
python-telegram-bot>=22.0

sounddevice==0.5.1
numpy==2.2.5
openai-agents[voice]
pydub==0.25.1